Refactor PoP base classes and register concrete implementations
