+++ pkgs/base/swarmauri_base/pop/__init__.py
@@ -1,11 +1,14 @@
 """Base helpers for proof-of-possession implementations."""
 
-from .base import PopSignerBase, PopVerifierBase, RequestContext
+from .PopSigningBase import PopSigningBase, PopSignerBase, RequestContext
+from .PopVerifierMixin import PopVerifierBase, PopVerifierMixin
 from .binding import make_cnf_cose, make_cnf_jkt, make_cnf_x5t, normalize_cnf
 from .util import canon_htm_htu, sha256_b64u
 
 __all__ = [
+    "PopSigningBase",
     "PopSignerBase",
+    "PopVerifierMixin",
     "PopVerifierBase",
     "RequestContext",
     "canon_htm_htu",
